+++ src/bacnet/basic/object/nc.c
@@ -847,6 +847,13 @@ int Notification_Class_Add_List_Element(BACNET_LIST_ELEMENT_DATA *list_element)
         if (len > 0) {
             new_element_count++;
             application_data_len -= len;
+            if (new_element_count >= NC_MAX_RECIPIENTS) {
+                list_element->first_failed_element_number = new_element_count;
+                list_element->error_class = ERROR_CLASS_RESOURCES;
+                list_element->error_code =
+                    ERROR_CODE_NO_SPACE_TO_ADD_LIST_ELEMENT;
+                return BACNET_STATUS_ERROR;
+            }
         } else {
             list_element->first_failed_element_number = new_element_count;
             list_element->error_class = ERROR_CLASS_PROPERTY;
@@ -1010,10 +1017,17 @@ int Notification_Class_Remove_List_Element(
         if (len > 0) {
             remove_element_count++;
             application_data_len -= len;
+            if (remove_element_count >= NC_MAX_RECIPIENTS) {
+                list_element->first_failed_element_number =
+                    remove_element_count;
+                list_element->error_class = ERROR_CLASS_SERVICES;
+                list_element->error_code = ERROR_CODE_LIST_ELEMENT_NOT_FOUND;
+                return BACNET_STATUS_ERROR;
+            }
         } else {
             list_element->first_failed_element_number = remove_element_count;
             list_element->error_class = ERROR_CLASS_PROPERTY;
-            list_element->error_code = ERROR_CODE_INVALID_DATA_ENCODING;
+            list_element->error_code = ERROR_CODE_INVALID_DATA_TYPE;
             return BACNET_STATUS_ERROR;
         }
     }