+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The git repositories are hosted at the following sites:
 
 ### Security
 
+* Secured Life Safety Point and Zone Read_Property of accepted-modes property
+  buffer overflow. (#1354)
 * Secured Notification Class object AddListElement and RemoveListElement stack
   based buffer overflow. (#1353)
 * Secured Timer object State_Change_Values property self-reference that