+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The git repositories are hosted at the following sites:
 
 ### Security
 
+* Secured Loop object internal Read_Property function buffer overflow. (#1355)
 * Secured Life Safety Point and Zone Read_Property of accepted-modes property
   buffer overflow. (#1354)
 * Secured Notification Class object AddListElement and RemoveListElement stack